applications, stats: fix cast widths and modernize logic

+++ src/applications/model/bulk-send-application.cc
@@ -173,7 +173,8 @@ BulkSendApplication::SendData(const Address& from, const Address& to)
         }
 
         int actual = m_socket->Send(packet);
-        if ((unsigned)actual == toSend)
+
+        if (static_cast<uint64_t>(actual) == toSend)
         {
             m_totBytes += actual;
             m_txTrace(packet);
@@ -194,9 +195,12 @@ BulkSendApplication::SendData(const Address& from, const Address& to)
             // a quantity less than the packet size.  Split the packet
             // into two, trace the sent packet, save the unsent packet
             NS_LOG_DEBUG("Packet size: " << packet->GetSize() << "; sent: " << actual
-                                         << "; fragment saved: " << toSend - (unsigned)actual);
+                                         << "; fragment saved: "
+                                         << toSend - static_cast<uint64_t>(actual));
             Ptr<Packet> sent = packet->CreateFragment(0, actual);
-            Ptr<Packet> unsent = packet->CreateFragment(actual, (toSend - (unsigned)actual));
+
+            Ptr<Packet> unsent =
+                packet->CreateFragment(actual, toSend - static_cast<uint64_t>(actual));
             m_totBytes += actual;
             m_txTrace(sent);
             m_unsentPacket = unsent;
